+++ README.md
@@ -60,41 +60,41 @@ public class Child {
 
 }
 ```
-在数据入库或者前端页面展示时可能需要对相应的敏感字段进行脱敏处理，你只需要在敏感数据字段上标记相应类型的敏感注解，例如`@ChineseNameSensitive`、`@EmailSensitive`等注解，如果你的数据字段是需要级联脱敏的对象，你只需要在该字段上标注`@CascadeSensitive`注解，最后调用以下方法即可擦除对象中的所有敏感信息然后返回一个新的Child对象，不会改变原对象中的域值。
+只需要在敏感数据字段上标记相应类型的敏感注解，例如`@ChineseNameSensitive`、`@EmailSensitive`等注解，如果你的数据字段是需要级联脱敏的对象，你只需要在该字段上标注`@CascadeSensitive`注解，最后调用以下方法即可擦除对象中的所有敏感信息然后返回一个新的Child对象。
 ```java
 Child child = Sensitive.desensitize(new Child());
 ```
 ### 值脱敏
 可能你的敏感信息是一个字符串类型的值或者是一个`Collection`、`Array`、`Map`之类的值，同样擦除它们的敏感信息也很简单：
 ```java
-private static void desensitizeValue(){
+static void desensitizeValue(){
 
-    // 单个值
-    System.out.println("值脱敏：" + Sensitive.desensitize("123456@qq.com", new TypeToken<@EmailSensitive String>() {
+    // String
+    System.out.printf("字符串脱敏: %s%n", Sensitive.desensitize("123456@qq.com", new TypeToken<@EmailSensitive String>() {
     }));
 
     // Collection
-    System.out.println("集合值脱敏：" + Sensitive.desensitize(Stream.of("123456@qq.com", "1234567@qq.com", "1234568@qq.com").collect(Collectors.toList()),
+    System.out.printf("集合脱敏: %s%n", Sensitive.desensitize(Stream.of("123456@qq.com", "1234567@qq.com", "1234568@qq.com").collect(Collectors.toList()),
             new TypeToken<List<@EmailSensitive String>>() {
             }));
 
     // Array
-    System.out.println("数组值脱敏：" + Arrays.toString(Sensitive.desensitize(new String[]{"123456@qq.com", "1234567@qq.com", "12345678@qq.com"},
+    System.out.printf("数组脱敏: %s%n", Arrays.toString(Sensitive.desensitize(new String[]{"123456@qq.com", "1234567@qq.com", "12345678@qq.com"},
             new TypeToken<@EmailSensitive String[]>() {
             })));
 
     // Map
-    System.out.println("Map值脱敏：" + Sensitive.desensitize(Stream.of("张三", "李四", "小明").collect(Collectors.toMap(s -> s, s -> "123456@qq.com")),
+    System.out.printf("Map脱敏: %s%n", Sensitive.desensitize(Stream.of("张三", "李四", "小明").collect(Collectors.toMap(s -> s, s -> "123456@qq.com")),
             new TypeToken<Map<@ChineseNameSensitive String, @EmailSensitive String>>() {
             }));
 }
 ```
 在上面的例子中通过`TypeToken`构造需要脱敏对象的类型以便我们能够准确的捕获被脱敏对象的实际类型和相应的敏感注解。
-这里有一个很重要的地方需要我们格外的关注：**由于jdk在解析注解时的bug导致无法正确的获取嵌套类上的注解，因此对于值脱敏时，我们需要将脱敏代码放到一个静态方法或者是静态块中，这样运行时才能正确的获取脱敏对象上的注解。**
+**这里有一个很重要的地方需要我们格外的关注：由于jdk在解析注解时的bug导致无法正确的获取嵌套类上的注解，TypeToken不能在实例方法、实例代码块中初始化同时也不能作为成员变量初始化，必须在静态方法、静态代码块中初始化或者作为静态变量初始化，这样运行时才能正确的获取脱敏对象上的注解。**
 有关这个bug的详情可以参考这个链接[why-annotation-on-generic-type-argument-is-not-visible-for-nested-type](http://stackoverflow.com/questions/39952812/why-annotation-on-generic-type-argument-is-not-visible-for-nested-type)
 # 例子
 1. [一个需要脱敏的复杂对象](https://github.com/Allurx/desensitization/blob/master/src/test/java/red/zyc/desensitization/model/Child.java)
-2. [测试用例](https://github.com/Allurx/desensitization/blob/master/src/test/java/red/zyc/desensitization/Example.java)
+2. [测试用例](https://github.com/Allurx/desensitization/blob/master/src/test/java/red/zyc/desensitization/DesensitizationTest.java)
 # 原理
 desensitization库是基于Java1.8新增的AnnotatedType这种新的类型体系来解析各种复杂数据结构中的脱敏注解，然后通过责任链这种设计模式完成数据脱敏处理的，要想完全理解其背后的实现原理需要对Java的Type体系和AnnotatedType体系有较为深刻的理解，可以参考以下几篇文章了解Java中的类型体系以及注解体系。
 * [Java Type](https://www.zyc.red/Java/Reflection/Type)
