+++ src/main/java/red/zyc/desensitization/Sensitive.java
@@ -46,9 +46,6 @@ import java.util.Optional;
  */
 public final class Sensitive {
 
-    private Sensitive() {
-    }
-
     /**
      * 对象内部域值脱敏，该方法不会改变原对象。
      *
@@ -62,7 +59,9 @@ public final class Sensitive {
     }
 
     /**
-     * 单个值脱敏，该方法不会改变原对象。
+     * 根据{@link TypeToken}脱敏对象，该方法不会改变原对象。
+     * <p><b>注意：{@link TypeToken}不能在实例方法、实例代码块中初始化同时也不能作为成员变量初始化，
+     * 必须在静态方法、静态代码块中初始化或者作为静态变量初始化。</b></p>
      *
      * @param target    目标对象
      * @param typeToken {@link TypeToken}
@@ -75,5 +74,8 @@ public final class Sensitive {
                 .map(TypeToken::getAnnotatedType)
                 .map(annotatedType -> TypeResolvers.resolve(target, annotatedType))
                 .orElse(target);
+    }
+
+    private Sensitive() {
     }
 }
